// api/src/entities/types/GarmentCategory.types.ts
import { GarmentSubCategoryMappedByCategoryLabel } from '../../utilities/constants/garmentSubCategory.constants';

export type GarmentCategoryLabelType = keyof typeof GarmentSubCategoryMappedByCategoryLabel;

export type GarmentSubCategoryByCategoryLabelType =
    | 'Dress'
    | 'Jacket'
    | 'Coat'
    | 'Bomber'
    | 'Vest'
    | 'Sweater'
    | 'Sweatshirt'
    | 'Down_jacket'
    | 'Shirt'
    | 'T-shirt'
    | 'Polo'
    | 'Tank_top'
    | 'Short_sleeve_shirt'
    | 'Sport_top'
    | 'Sleeveless_top'
    | 'Pants'
    | 'Skirt'
    | 'Shorts'
    | 'Leggings'
    | 'Jeans'
    | 'Bermuda_shorts'
    | 'Jogging_pants'
    | 'Overalls'
    | 'Cargo_pants'
    | 'Sneakers'
    | 'Sandals'
    | 'Boots'
    | 'Dress_shoes'
    | 'Loafers'
    | 'Espadrilles'
    | 'Flip-flops'
    | 'Ankle_boots'
    | 'Sports_shoes'
    | 'Hiking_shoes'
    | 'Flats';